+++ tests/unit/IntegrationRouter/App.vue
@@ -9,6 +9,7 @@
 
 import {container} from "../../../plugin";
 
+
 export default {
 	components: {WidgetModalContainer: container},
 	name: "App"